+++ glean_parser/translate.py
@@ -16,6 +16,7 @@ from typing import Any, Callable, Dict, Iterable, List
 
 from . import lint
 from . import parser
+from . import csharp
 from . import kotlin
 from . import markdown
 from . import metrics
@@ -46,6 +47,7 @@ class Outputter:
 
 
 OUTPUTTERS = {
+    "csharp": Outputter(csharp.output_csharp, ["*.cs"]),
     "kotlin": Outputter(kotlin.output_kotlin, ["*.kt"]),
     "markdown": Outputter(markdown.output_markdown),
     "swift": Outputter(swift.output_swift, ["*.swift"]),