+++ glean_parser/csharp.py
@@ -127,25 +127,6 @@ def output_csharp(
         ),
     )
 
-    # The object parameters to pass to constructors
-    extra_args = [
-        "allowed_extra_keys",
-        "bucket_count",
-        "category",
-        "disabled",
-        "histogram_type",
-        "include_client_id",
-        "send_if_empty",
-        "lifetime",
-        "memory_unit",
-        "name",
-        "range_max",
-        "range_min",
-        "reason_codes",
-        "send_in_pings",
-        "time_unit",
-    ]
-
     namespace = options.get("namespace", "GleanMetrics")
     glean_namespace = options.get("glean_namespace", "Mozilla.Glean")
 
@@ -158,7 +139,7 @@ def output_csharp(
                 template.render(
                     category_name=category_key,
                     objs=category_val,
-                    extra_args=extra_args,
+                    extra_args=util.extra_args,
                     namespace=namespace,
                     glean_namespace=glean_namespace,
                 )