+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 Unreleased
 ----------
 
+* Add support for generating C# code.
+
 1.24.0 (2020-06-30)
 -------------------